+++ lib/teams_connector/notification.rb
@@ -30,7 +30,7 @@ module TeamsConnector
         elsif TeamsConnector.configuration.method == :testing
           TeamsConnector.testing.perform_request channel, @template, content
         else
-          response = Net::HTTP.post(URI(url), content, { "Content-Type": "application/json" })
+          response = Net::HTTP.post(URI(url), content, { "Content-Type" => "application/json" })
           response.value
         end
       end