+++ lib/teams_connector/rspec.rb
@@ -1,0 +1,5 @@
+require "teams_connector/matchers"
+
+RSpec.configure do |config|
+  config.include TeamsConnector::Matchers
+end