+++ spec/teams_connector/notification/message_spec.rb
@@ -6,9 +6,8 @@ RSpec.describe TeamsConnector::Notification::Message do
       config.channel :other, "http://localhost"
       config.channel :default, "http://default"
       config.default = :default
+      config.method = :testing
     end
-    stub_request(:post, "localhost")
-    stub_request(:post, "default")
   end
 
   subject { TeamsConnector::Notification::Message.new(:test_card, "Summary") }
@@ -19,5 +18,12 @@ RSpec.describe TeamsConnector::Notification::Message do
     expect(subject).to have_attributes(channels: [:default])
     expect(subject).to have_attributes(summary: "Summary")
     expect(subject).to have_attributes(content: {})
+  end
+
+  it "sends a notification with the content" do
+    expect { subject.deliver_later }.to have_sent_notification_to(:default).with_template(:test_card).with { |content|
+      expect(content["summary"]).to eq "This is a test summary"
+      expect(content["sections"]).to include(hash_including("facts", "markdown" => true, "activityTitle" => "Quokka", "activitySubtitle" => "About the short-tailed scrub wallaby"))
+    }
   end
 end